【BUG】C++ Bson的toString输出Date类型和Timestamp类型，如果时间小于1970年会错误 SEQUOIADBMAINSTREAM-1068
git-svn-id: http://192.168.20.11/sequoiadb/trunk@20097 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/bson/bson-inl.h
@@ -885,7 +885,7 @@ namespace bson {
             Date_t date = timestampTime () ;
             unsigned int inc = timestampInc () ;
             char buffer[128] ;
-            time_t timer = (time_t)(date.millis/1000) ;
+            time_t timer = (time_t)(((long long)date.millis)/1000) ;
             struct tm psr ;
             local_time ( &timer, &psr ) ;
             memset ( buffer, 0, 128 ) ;
